Move ModSounds to content folder
