+++ src/main/java/com/fredtargaryen/rocketsquids/ModRocketSquids.java
@@ -3,7 +3,6 @@
 package com.fredtargaryen.rocketsquids;
 
 import com.fredtargaryen.rocketsquids.client.event.ModClientHandler;
-import com.fredtargaryen.rocketsquids.client.particle.SquidFireworkParticle;
 import com.fredtargaryen.rocketsquids.config.Config;
 import com.fredtargaryen.rocketsquids.config.GeneralConfig;
 import com.fredtargaryen.rocketsquids.content.*;
@@ -16,9 +15,6 @@ import com.fredtargaryen.rocketsquids.content.cap.item.squeleporter.Squeleporter
 import com.fredtargaryen.rocketsquids.content.entity.BabyRocketSquidEntity;
 import com.fredtargaryen.rocketsquids.content.entity.RocketSquidEntity;
 import com.fredtargaryen.rocketsquids.network.MessageHandler;
-import net.minecraft.client.Minecraft;
-import net.minecraft.core.particles.ParticleType;
-import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.world.entity.Entity;
@@ -26,7 +22,6 @@ import net.minecraft.world.entity.SpawnPlacements;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.biome.MobSpawnSettings;
 import net.minecraft.world.level.levelgen.Heightmap;
-import net.minecraftforge.client.event.RegisterParticleProvidersEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
@@ -38,9 +33,6 @@ import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLPaths;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegistryObject;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
