+++ src/main/java/com/fredtargaryen/rocketsquids/content/ModSounds.java
@@ -1,6 +1,6 @@
 // Copyright 2016-2022, 2025-2026 FredTargaryen and contributors
 // See README.md for full copyright notice and contributor info
-package com.fredtargaryen.rocketsquids;
+package com.fredtargaryen.rocketsquids.content;
 
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
