+++ src/main/java/com/fredtargaryen/rocketsquids/content/block/StatueBlock.java
@@ -1,6 +1,6 @@
 package com.fredtargaryen.rocketsquids.content.block;
 
-import com.fredtargaryen.rocketsquids.ModSounds;
+import com.fredtargaryen.rocketsquids.content.ModSounds;
 import com.fredtargaryen.rocketsquids.content.ModItems;
 import com.fredtargaryen.rocketsquids.content.worldgen.StatueData;
 import net.minecraft.core.BlockPos;
